@@ -1,9 +1,14 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import './index.css';
+import AppContextProvider from './context/AppContext';
 import App from './App';
-import * as serviceWorker from './serviceWorker';
+import AppNew from './AppNew';
 
-
-ReactDOM.render(<App />, document.getElementById('root'));
-serviceWorker.unregister();
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
+    <AppContextProvider >
+        {/* <App /> */}
+        <AppNew />
+    </AppContextProvider>
+);
