@@ -1,29 +1,37 @@
-import React, { Component } from "react";
+import React, { useContext } from "react";
 import "./SkillCard.css";
+import { AppContext } from "../context/AppContext";
 
 
-class SkillCard extends Component {
-  render() {
-    const { skill } = this.props;
-    return (
-      <div className="card" style={{
-        backgroundImage :`linear-gradient(0deg ,
-          ${skill.content.backgroundcolor2}  ${skill.content.score}%
-          , ${skill.content.backgroundcolor1}  ${100 - skill.content.score}%)` }} >
+
+const SkillCard = ({index}) => {
+
+  const [state] = useContext(AppContext);
+
+  const skillLanguages = state?.skillLanguages || {};
+  const items = skillLanguages?.items || [];
+  const skill = items[index];
 
 
-        <div className="image-wrapper">
-          <img src={skill.content.image} alt="" />
-        </div>
-        <div className="skill-title-wrapper">
-          <h4>{skill.content.title}</h4>
-        </div>
+  return (
+    <div className="card" style={{
+      backgroundImage: `linear-gradient(0deg ,
+        ${skill?.content?.backgroundColor2}  ${skill?.content?.score}%
+        , ${skill?.content?.backgroundColor1}  ${100 - skill?.content?.score}%)`
+    }} >
 
-        
 
+      <div className="image-wrapper">
+        <img src={skill?.content?.image} alt="" />
       </div>
-    );
-  }
+      <div className="skill-title-wrapper">
+        <h4>{skill?.content?.title}</h4>
+      </div>
+
+
+
+    </div>
+  );
 }
 
 export default SkillCard;
