@@ -1,6 +1,6 @@
 import React, { Component } from "react";
-import Fullpage from "../components/Fullpage";
-import skilllanguagessss from "./EnginieeringToools.json";
+import FullPage from "../components/FullPage";
+import skillLanguages from "./EnginieeringToools.json";
 import data from "../data.json";
 import SkillCard from '../components/SkillCard';
 import DownIcon from "../components/DownIcon";
@@ -10,23 +10,23 @@ import "./EnginieeringToools.css";
 class EnginieeringToools extends Component {
   render() {
     return (
-      <div name="engtools">      <Fullpage className="EnginieeringToools">
+      <div name="engtools">      <FullPage className="EnginieeringToools">
        
-        <h3  style={{color: "balck"}} className="EnginieeringToools" >{skilllanguagessss.title}</h3>
+        <h3  style={{color: "balck"}} className="EnginieeringToools" >{skillLanguages.title}</h3>
         <div className="cards-wrapper">
-          {skilllanguagessss.items.map(eachSkill => {
-            eachSkill.content['backgroundcolor1']  = skilllanguagessss.backgroundcolor1
-            eachSkill.content['backgroundcolor2'] = skilllanguagessss.backgroundcolor2
+          {skillLanguages.items.map(eachSkill => {
+            eachSkill.content['backgroundColor1']  = skillLanguages.backgroundColor1
+            eachSkill.content['backgroundColor2'] = skillLanguages.backgroundColor2
             return (
               <SkillCard className="EnginieeringTooolscard" skill={eachSkill}/>
             );
           })}
         </div>
-      </Fullpage>
+      </FullPage>
       
       <Link
           activeClass="active"
-          to="maincontentent"
+          to="mainContentSection"
           spy={true}
           smooth={true}
           offset={0}
