@@ -1,5 +1,5 @@
 import React, { Component } from "react";
-import Fullpage from "../components/Fullpage";
+import FullPage from "../components/FullPage";
 import data from "../data.json";
 import SkillCard from '../components/SkillCard';
 import DownIcon from "../components/DownIcon";
@@ -9,23 +9,23 @@ import "./SkillSection.css";
 class AboutSection extends Component {
   render() {
     return (
-      <div name="otherss">      <Fullpage className="third">
+      <div name="otherss">      <FullPage className="third">
        
         <h3>{data.sections[1].title}</h3>
         <div className="cards-wrapper">
           {data.sections[1].items.map(eachSkill => {
-            eachSkill.content['backgroundcolor1']  = data.sections[1].backgroundcolor1
-            eachSkill.content['backgroundcolor2'] = data.sections[1].backgroundcolor2
+            eachSkill.content['backgroundColor1']  = data.sections[1].backgroundColor1
+            eachSkill.content['backgroundColor2'] = data.sections[1].backgroundColor2
             return (
               <SkillCard skill={eachSkill}/>
             );
           })}
         </div>
-      </Fullpage>
+      </FullPage>
       
       <Link
           activeClass="active"
-          to="maincontentent"
+          to="mainContentSection"
           spy={true}
           smooth={true}
           offset={0}
