@@ -1,6 +1,6 @@
 import React, { Component } from "react";
-import Fullpage from "../components/Fullpage";
-import skilllanguagessss from "./Webservers.json";
+import FullPage from "../components/FullPage";
+import skillLanguages from "./Webservers.json";
 import data from "../data.json";
 import SkillCard from '../components/SkillCard';
 import DownIcon from "../components/DownIcon";
@@ -10,23 +10,23 @@ import "./Webservers.css";
 class Webservers extends Component {
   render() {
     return (
-      <div name="portfolioscrool">      <Fullpage className="Webservers">
+      <div name="portfolioScrool">      <FullPage className="Webservers">
        
-        <h3  style={{color: "balck"}} className="Webservers" >{skilllanguagessss.title}</h3>
+        <h3  style={{color: "balck"}} className="Webservers" >{skillLanguages.title}</h3>
         <div className="cards-wrapper">
-          {skilllanguagessss.items.map(eachSkill => {
-            eachSkill.content['backgroundcolor1']  = skilllanguagessss.backgroundcolor1
-            eachSkill.content['backgroundcolor2'] = skilllanguagessss.backgroundcolor2
+          {skillLanguages.items.map(eachSkill => {
+            eachSkill.content['backgroundColor1']  = skillLanguages.backgroundColor1
+            eachSkill.content['backgroundColor2'] = skillLanguages.backgroundColor2
             return (
               <SkillCard className="Webserverscard" skill={eachSkill}/>
             );
           })}
         </div>
-      </Fullpage>
+      </FullPage>
       
       <Link
           activeClass="active"
-          to="maincontentent"
+          to="mainContentSection"
           spy={true}
           smooth={true}
           offset={0}
